+++ projects/ngx-tools/src/directives/countdown.directive.ts
@@ -1,5 +1,5 @@
 import {Directive, ElementRef, inject, Input, LOCALE_ID, NgZone, OnChanges, SimpleChanges} from '@angular/core';
-import {objToArr, sortNumDesc} from "@juulsgaard/ts-tools";
+import {objToArr, sortNumDesc, Timespan} from "@juulsgaard/ts-tools";
 import {concat, EMPTY, endWith, fromEvent, interval, share, startWith, Subscription, takeWhile, timer} from "rxjs";
 import {distinctUntilChanged, filter, map, tap} from "rxjs/operators";
 import {Dispose} from "../decorators";
@@ -102,14 +102,14 @@ export class CountdownDirective implements OnChanges {
     }
 
     const dateDelta = delta - this.dateFormatThreshold;
-    const dateDelay$ = dateDelta <= 0 ? EMPTY : timer(Math.max(0, dateDelta + 1)).pipe(
+    const dateDelay$ = dateDelta <= 0 ? EMPTY : timer(Math.max(0, Timespan.seconds(dateDelta + 1))).pipe(
       tap({subscribe: () => this.renderDate()}),
       filter(() => false)
     );
 
     const maxTimeDelay = Math.max(0, this.dateFormatThreshold - this.timeFormatThreshold - 1);
     const timeDelta = Math.min(delta - this.timeFormatThreshold, maxTimeDelay);
-    const timeDelay$ = timeDelta <= 0 ? EMPTY : timer(timeDelta).pipe(
+    const timeDelay$ = timeDelta <= 0 ? EMPTY : timer(Timespan.seconds(timeDelta)).pipe(
       tap({subscribe: () => this.renderTime()}),
       filter(() => false)
     );
@@ -212,20 +212,23 @@ export class CountdownDirective implements OnChanges {
   }
 
   private renderDate() {
-    this.element.classList.remove(['ngx-counting', 'ngx-time']);
-    this.element.classList.add(['ngx-date']);
+    this.element.classList.remove('ngx-counting');
+    this.element.classList.remove('ngx-time');
+    this.element.classList.add('ngx-date');
     this.dateTimeNode.innerText = formatDate(this.endTime, 'short', this.locale);
   }
 
   private renderTime() {
-    this.element.classList.remove(['ngx-counting', 'ngx-date']);
-    this.element.classList.add(['ngx-time']);
+    this.element.classList.remove('ngx-counting');
+    this.element.classList.remove('ngx-date');
+    this.element.classList.add('ngx-time');
     this.dateTimeNode.innerText = formatDate(this.endTime, 'shortTime', this.locale);
   }
 
   private countdownStarted() {
-    this.element.classList.remove(['ngx-time', 'ngx-date']);
-    this.element.classList.add(['ngx-counting']);
+    this.element.classList.remove('ngx-time');
+    this.element.classList.remove('ngx-date');
+    this.element.classList.add('ngx-counting');
   }
 
   appliedClasses?: string[];